+++ janusgraph-cql/src/test/java/org/janusgraph/diskstorage/cql/CachingCQLStoreManager.java
@@ -41,9 +41,6 @@ public class CachingCQLStoreManager extends CQLStoreManager {
             sessions.values().forEach(CqlSession::close);
             sessions.clear();
         }
-        if(this.executorService != null){
-            this.executorService.shutdownNow();
-        }
     }
 
 }